# openapi-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/customers": {
      "post": {
        "operationId": "CustomersController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCustomerDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Cliente criado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Customer"
                }
              }
            }
          },
          "400": {
            "description": "Dados de entrada inválidos"
          },
          "409": {
            "description": "Cliente com mesmo CNPJ/CPF já existe"
          }
        },
        "summary": "Criar novo cliente",
        "tags": [
          "Customers"
        ]
      },
      "get": {
        "operationId": "CustomersController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Lista de clientes retornada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Customer"
                  }
                }
              }
            }
          }
        },
        "summary": "Listar todos os clientes",
        "tags": [
          "Customers"
        ]
      }
    },
    "/api/customers/{id}": {
      "get": {
        "operationId": "CustomersController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID do cliente (número)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Cliente encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Customer"
                }
              }
            }
          },
          "404": {
            "description": "Cliente não encontrado"
          }
        },
        "summary": "Obter cliente por ID",
        "tags": [
          "Customers"
        ]
      },
      "patch": {
        "operationId": "CustomersController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID do cliente (número)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCustomerDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Cliente atualizado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Customer"
                }
              }
            }
          },
          "400": {
            "description": "Dados de entrada inválidos"
          },
          "404": {
            "description": "Cliente não encontrado"
          }
        },
        "summary": "Atualizar cliente por ID",
        "tags": [
          "Customers"
        ]
      },
      "delete": {
        "operationId": "CustomersController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID do cliente (número)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Cliente removido com sucesso"
          },
          "404": {
            "description": "Cliente não encontrado"
          }
        },
        "summary": "Remover cliente por ID",
        "tags": [
          "Customers"
        ]
      }
    },
    "/api/customers/by-document/{cnpjCpf}": {
      "get": {
        "operationId": "CustomersController_findByDocument",
        "parameters": [
          {
            "name": "cnpjCpf",
            "required": true,
            "in": "path",
            "description": "CNPJ/CPF do cliente",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Cliente encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Customer"
                }
              }
            }
          },
          "404": {
            "description": "Cliente não encontrado"
          }
        },
        "summary": "Obter cliente por CNPJ/CPF",
        "tags": [
          "Customers"
        ]
      }
    },
    "/api/countries": {
      "post": {
        "operationId": "CountriesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCountryDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "País criado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Country"
                }
              }
            }
          },
          "400": {
            "description": "Dados de entrada inválidos"
          },
          "409": {
            "description": "País com mesmo código ou sigla já existe"
          }
        },
        "summary": "Criar novo país",
        "tags": [
          "Countries"
        ]
      },
      "get": {
        "operationId": "CountriesController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Lista de países retornada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Country"
                  }
                }
              }
            }
          }
        },
        "summary": "Listar todos os países",
        "tags": [
          "Countries"
        ]
      }
    },
    "/api/countries/code/{code}": {
      "get": {
        "operationId": "CountriesController_findByCode",
        "parameters": [
          {
            "name": "code",
            "required": true,
            "in": "path",
            "description": "Código do país",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "País encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Country"
                }
              }
            }
          },
          "404": {
            "description": "País não encontrado"
          }
        },
        "summary": "Obter país por código",
        "tags": [
          "Countries"
        ]
      }
    },
    "/api/countries/sigla/{sigla}": {
      "get": {
        "operationId": "CountriesController_findBySigla",
        "parameters": [
          {
            "name": "sigla",
            "required": true,
            "in": "path",
            "description": "Sigla do país",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "País encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Country"
                }
              }
            }
          },
          "404": {
            "description": "País não encontrado"
          }
        },
        "summary": "Obter país por sigla",
        "tags": [
          "Countries"
        ]
      }
    },
    "/api/countries/{id}": {
      "get": {
        "operationId": "CountriesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID do país",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "País encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Country"
                }
              }
            }
          },
          "404": {
            "description": "País não encontrado"
          }
        },
        "summary": "Obter país por ID",
        "tags": [
          "Countries"
        ]
      },
      "patch": {
        "operationId": "CountriesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID do país",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCountryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "País atualizado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Country"
                }
              }
            }
          },
          "400": {
            "description": "Dados de entrada inválidos"
          },
          "404": {
            "description": "País não encontrado"
          },
          "409": {
            "description": "País com mesmo código ou sigla já existe"
          }
        },
        "summary": "Atualizar país por ID",
        "tags": [
          "Countries"
        ]
      },
      "delete": {
        "operationId": "CountriesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID do país",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "País removido com sucesso"
          },
          "404": {
            "description": "País não encontrado"
          },
          "409": {
            "description": "Não é possível excluir o país pois ele possui estados vinculados"
          }
        },
        "summary": "Remover país por ID",
        "tags": [
          "Countries"
        ]
      }
    },
...
  },
  "info": {
    "title": "Serveon API",
    "description": "API RESTful para o sistema Serveon",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateCustomerDto": {
        "type": "object",
        "properties": {
          "cnpjCpf": {
            "type": "string",
            "description": "CNPJ, CPF ou número de documento internacional do cliente",
            "example": "12345678901234"
          },
          "tipo": {
            "type": "string",
            "description": "Tipo de cliente: F para Pessoa Física, J para Pessoa Jurídica",
            "example": "J",
            "enum": [
              "F",
              "J"
            ]
          },
          "isEstrangeiro": {
            "type": "boolean",
            "description": "Indica se é um cliente estrangeiro",
            "example": false,
            "default": false
          },
          "tipoDocumento": {
            "type": "string",
            "description": "Tipo de documento para clientes estrangeiros",
            "example": "passport",
            "enum": [
              "passport",
              "tax_id",
              "national_id",
              "other"
            ]
          },
          "razaoSocial": {
            "type": "string",
            "description": "Razão Social ou Nome Completo",
            "example": "Empresa XYZ Ltda"
          },
          "nomeFantasia": {
            "type": "string",
            "description": "Nome Fantasia (opcional)",
            "example": "XYZ Tecnologia"
          },
          "inscricaoEstadual": {
            "type": "string",
            "description": "Inscrição Estadual ou dados adicionais do documento estrangeiro",
            "example": "123456789"
          },
          "inscricaoMunicipal": {
            "type": "string",
            "description": "Inscrição Municipal (opcional)",
            "example": "123456"
          },
          "paisId": {
            "type": "number",
            "description": "ID do país",
            "example": 1
          },
          "estadoId": {
            "type": "number",
            "description": "ID do estado/província",
            "example": 1
          },
          "cidadeId": {
            "type": "number",
            "description": "ID da cidade no sistema",
            "example": 1
          },
          "endereco": {
            "type": "string",
            "description": "Endereço",
            "example": "Av. Brasil"
          },
          "numero": {
            "type": "string",
            "description": "Número do endereço (opcional)",
            "example": "123"
          },
          "complemento": {
            "type": "string",
            "description": "Complemento do endereço (opcional)",
            "example": "Sala 101"
          },
          "bairro": {
            "type": "string",
            "description": "Bairro (opcional)",
            "example": "Centro"
          },
          "cep": {
            "type": "string",
            "description": "CEP ou Código Postal",
            "example": "12345678"
          },
          "telefone": {
            "type": "string",
            "description": "Telefone (opcional)",
            "example": "11987654321"
          },
          "email": {
            "type": "string",
            "description": "Email (opcional)",
            "example": "contato@empresa.com"
          },
          "ativo": {
            "type": "boolean",
            "description": "Status ativo/inativo do cliente",
            "example": true,
            "default": true
          },
          "isDestinatario": {
            "type": "boolean",
            "description": "Se o cliente também é um destinatário",
            "example": true,
            "default": true
          },
          "destinatariosIds": {
            "description": "Lista de IDs de destinatários associados (quando o cliente não é o destinatário)",
            "example": [
              1
            ],
            "type": "array",
            "items": {
              "type": "number"
            }
          }
        },
        "required": [
          "cnpjCpf",
          "tipo",
          "isEstrangeiro",
          "razaoSocial",
          "ativo",
          "isDestinatario"
        ]
      },
      "Customer": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "ID único do cliente",
            "example": 1
          },
          "cnpjCpf": {
            "type": "string",
            "description": "CNPJ, CPF ou documento internacional do cliente",
            "example": "12345678901234"
          },
          "tipo": {
            "type": "string",
            "description": "Tipo de cliente: F para Pessoa Física, J para Pessoa Jurídica",
            "example": "J",
            "enum": [
              "F",
              "J"
            ]
          },
          "isEstrangeiro": {
            "type": "boolean",
            "description": "Indica se é um cliente estrangeiro",
            "example": false
          },
          "tipoDocumento": {
            "type": "string",
            "description": "Tipo de documento para clientes estrangeiros",
            "example": "passport"
          },
          "razaoSocial": {
            "type": "string",
            "description": "Razão Social ou Nome Completo",
            "example": "Empresa XYZ Ltda"
          },
          "nomeFantasia": {
            "type": "string",
            "description": "Nome Fantasia",
            "example": "XYZ Tecnologia"
          },
          "inscricaoEstadual": {
            "type": "string",
            "description": "Inscrição Estadual ou informações complementares do documento",
            "example": "123456789"
          },
          "inscricaoMunicipal": {
            "type": "string",
            "description": "Inscrição Municipal ou registro comercial internacional",
            "example": "123456"
          },
          "paisId": {
            "type": "number",
            "description": "ID do país",
            "example": 1
          },
          "paisNome": {
            "type": "string",
            "description": "Nome do país",
            "example": "Brasil"
          },
          "estadoId": {
            "type": "number",
            "description": "ID do estado/província",
            "example": 1
          },
          "endereco": {
            "type": "string",
            "description": "Endereço",
            "example": "Av. Brasil"
          },
          "numero": {
            "type": "string",
            "description": "Número do endereço",
            "example": "123"
          },
          "complemento": {
            "type": "string",
            "description": "Complemento do endereço",
            "example": "Sala 101"
          },
          "bairro": {
            "type": "string",
            "description": "Bairro",
            "example": "Centro"
          },
          "cidadeId": {
            "type": "number",
            "description": "ID da cidade",
            "example": 1
          },
          "cep": {
            "type": "string",
            "description": "CEP ou código postal",
            "example": "12345678"
          },
          "telefone": {
            "type": "string",
            "description": "Telefone",
            "example": "11987654321"
          },
          "email": {
            "type": "string",
            "description": "Email",
            "example": "contato@empresa.com"
          },
          "ativo": {
            "type": "boolean",
            "description": "Status ativo/inativo do cliente",
            "example": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Data de criação",
            "example": "2023-01-01T00:00:00.000Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Data da última atualização",
            "example": "2023-01-01T00:00:00.000Z"
          },
          "cidadeNome": {
            "type": "string",
            "description": "Nome da cidade",
            "example": "São Paulo"
          },
          "estadoNome": {
            "type": "string",
            "description": "Nome do estado",
            "example": "São Paulo"
          },
          "uf": {
            "type": "string",
            "description": "UF do estado",
            "example": "SP"
          },
          "destinatariosIds": {
            "description": "Lista de IDs de destinatários associados ao cliente",
            "example": [
              1,
              2,
              3
            ],
            "type": "array",
            "items": {
              "type": "number"
            }
          },
          "isDestinatario": {
            "type": "boolean",
            "description": "Se o cliente é também um destinatário",
            "example": true
          }
        },
        "required": [
          "id",
          "cnpjCpf",
          "tipo",
          "isEstrangeiro",
          "razaoSocial",
          "nomeFantasia",
          "inscricaoEstadual",
          "inscricaoMunicipal",
          "paisId",
          "paisNome",
          "estadoId",
          "endereco",
          "numero",
          "complemento",
          "bairro",
          "cidadeId",
          "cep",
          "telefone",
          "email",
          "ativo",
          "createdAt",
          "updatedAt",
          "cidadeNome",
          "estadoNome",
          "uf",
          "isDestinatario"
        ]
      },
      "UpdateCustomerDto": {
        "type": "object",
        "properties": {
          "ativo": {
            "type": "boolean",
            "description": "Status ativo/inativo do cliente",
            "example": true
          },
          "isDestinatario": {
            "type": "boolean",
            "description": "Se o cliente também é um destinatário",
            "example": true
          },
          "destinatariosIds": {
            "description": "Lista de IDs de destinatários associados (quando o cliente não é o destinatário)",
            "example": [
              1,
              2,
              3
            ],
            "type": "array",
            "items": {
              "type": "number"
            }
          }
        }
      },
      "CreateCountryDto": {
        "type": "object",
        "properties": {
          "nome": {
            "type": "string",
            "description": "Nome do país",
            "example": "Brasil"
          },
          "codigo": {
            "type": "string",
            "description": "Código de chamada internacional do país",
            "example": "55"
          },
          "sigla": {
            "type": "string",
            "description": "Sigla do país (2 caracteres)",
            "example": "BR"
          }
        },
        "required": [
          "nome",
          "codigo",
          "sigla"
        ]
      },
      "Country": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "ID único do país",
            "example": 1
          },
          "nome": {
            "type": "string",
            "description": "Nome do país",
            "example": "Brasil"
          },
          "codigo": {
            "type": "string",
            "description": "Código de chamada internacional do país",
            "example": "55"
          },
          "sigla": {
            "type": "string",
            "description": "Sigla do país",
            "example": "BR"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Data de criação do registro",
            "example": "2023-01-15T12:00:00.000Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Data da última atualização do registro",
            "example": "2023-01-15T12:00:00.000Z"
          }
        },
        "required": [
          "id",
          "nome",
          "codigo",
          "sigla",
          "createdAt",
          "updatedAt"
        ]
      },
      "UpdateCountryDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}